booleanDTO and so on

+++ author-book-ORM/src/main/java/com/example/demo/models/controller/BookController.java
@@ -1,30 +1,24 @@
 package com.example.demo.models.controller;
 
-import com.example.demo.models.DTO.AuthorDTO;
-import com.example.demo.models.DTO.AuthorRequest;
 import com.example.demo.models.DTO.BookDTO;
-import com.example.demo.models.DTO.TagDTO;
+import com.example.demo.models.exceptions.BookRejectionException;
 import com.example.demo.models.gateway.BookServiceGateway;
-import com.example.demo.models.service.AuthorService;
+import com.example.demo.models.gateway.HttpBookServiceGateway;
 import com.example.demo.models.service.BookService;
-import com.example.demo.models.entity.Book;
 import com.example.demo.models.DTO.BookRequest;
-import com.example.demo.models.DTO.TagRequest;
-import com.example.demo.models.service.TagService;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.client.RestTemplate;
 
 import java.util.List;
-import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
 
 @RestController
 @RequestMapping("/api/books")
@@ -46,17 +40,39 @@ public class BookController {
 
   @PostMapping(path = "/add")
   public BookDTO createBook(@Valid @RequestBody BookRequest book) {
+    if (!bookServiceGateway.checkBookExists(
+                new BookDTO(null, book.getAuthorID(), book.getTitle(), null),
+                UUID.randomUUID().toString()).value()) {
+      throw new BookRejectionException("Such a book wasn't verified", new RuntimeException());
+    }
     return bookService.add(book);
   }
 
   @PutMapping("/{id}")
   public BookDTO updateBook(@NotNull @PathVariable("id") Long id,
                          @Valid @RequestBody BookRequest update) {
+    Optional<BookDTO> bookDTO = bookService.getById(id, false);
+    if (!bookServiceGateway.checkBookExists(
+                new BookDTO(
+                    null,
+                    update.getAuthorID() == null ? bookDTO.get().authorID() : update.getAuthorID(),
+                    update.getTitle() == null ? bookDTO.get().title() : update.getTitle(),
+                    null),
+                UUID.randomUUID().toString()).value()) {
+      throw new BookRejectionException("Such a book wasn't verified or created", new RuntimeException());
+    }
     return bookService.update(id, update);
   }
 
   @DeleteMapping("/{id}")
   public void deleteBook(@NotNull @PathVariable("id") Long id) {
+    Optional<BookDTO> bookDTO = bookService.getById(id, false);
+    if (!bookServiceGateway.checkBookExists(
+            new BookDTO(null, bookDTO.get().authorID(), bookDTO.get().title(), null),
+            UUID.randomUUID().toString()).value()) {
+      throw new BookRejectionException("Such a book wasn't verified or created", new RuntimeException());
+    }
+
     bookService.delete(id);
   }
 
@@ -71,5 +87,10 @@ public class BookController {
     } catch (NumberFormatException e) {
       return bookService.getWithTag(tag);
     }
+  }
+
+  @ExceptionHandler
+  public ResponseEntity<Exception> nullPointerException(NullPointerException ex) {
+    return ResponseEntity.status(422).body(new Exception(ex.getMessage()));
   }
 }
